BUILD-13322: fix notary key map failure (#158)
This commit fixes building repos for Mongo versions in the form of 5.0.0-rc0-29-gaebf6fb.
Versions in this form were  not mapped to their correct notary keys and causing build failures.
The primary cause was bond.StableReleaseSeries() not existing for new MongoDB versions,
so the logic was simplified using just LTS() and Series(). Test cases have been added
as well.