EVG-18207: Only explicitly exit on SIGINT, SIGKILL, and SIGHUP.
We only need to do this handling so that we can make a best effort
at reading to the end of the file before we close, so it only
makes sense to handles signals with a default "Terminate" action
that users would likely use to control the file tailing.

None of the other signals look likely to be used for control by
external processes, and the remaining signals that are ignored
or have other actions probably shouldn't be specially handled by
us.

This fixes an issue where we were killing our process on SIGURG.